refactor: rename formatter and logger files for consistency
