+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+  "files.exclude": {
+    "**/*.log*": false
+  },
+  "search.exclude": {
+    "**/*.log*": false
+  },
+  "files.watcherExclude": {
+    "**/*.log*": false
+  }
+}
