+++ src/modules/formatters/implementations/ansi.formatter.ts
@@ -1,5 +1,5 @@
 import type { HandlerError } from "../../../core/handler-error";
-import { ErrorFormatter } from "../base-formatter";
+import { ErrorFormatter } from "../base.formatter";
 
 export const ANSI_COLORS = {
   black: "\u001B[30m",
