+++ src/modules/index.ts
@@ -1,10 +1,10 @@
 // Formatters
-export * from "./formatters/base-formatter";
-export * from "./formatters/implementations/ansi-formatter";
-export * from "./formatters/implementations/html-formatter";
-export * from "./formatters/implementations/json-formatter";
-export * from "./formatters/implementations/text-formatter";
+export * from "./formatters/base.formatter";
+export * from "./formatters/implementations/ansi.formatter";
+export * from "./formatters/implementations/html.formatter";
+export * from "./formatters/implementations/json.formatter";
+export * from "./formatters/implementations/text.formatter";
 
 // Loggers
-export * from "./loggers/base-logger";
-export * from "./loggers/implementations/console-logger";
+export * from "./loggers/base.logger";
+export * from "./loggers/implementations/console.logger";
